+++ Tests/FilesTest.cs
@@ -44,10 +44,14 @@ namespace SharePointPnP.PowerShell.Tests
             }
         }
 
+        private const string SourceFolderWithFolders = "SourceFolderWithFoldersAndFilesAndEmptyFolder";
+        private const string SourceFolderName = "SourceFolder";
         private const string TargetFileName = "Testfile.txt";
         private const string TargetFileContents = "Some random file contents";
         private const string TargetCopyFolderName = "CopyDestination";
+        private const string EmptyFolderName = "EmptyFolder";
         private const string TargetFileNameWithAmpersand = "Test & file.txt";
+        private const string TargetFileNameWithHashtag = "Test & file.txt";
 
         [TestInitialize]
         public void Initialize()
@@ -87,10 +91,34 @@ namespace SharePointPnP.PowerShell.Tests
                     fileToUpload = folder.Files.Add(fci);
                     site1Ctx.Load(fileToUpload);
 
-                    site1Ctx.ExecuteQueryRetry();
+                    fci.Url = TargetFileNameWithHashtag;
+                    fci.Overwrite = true;
+                    fileToUpload = folder.Files.Add(fci);
+                    site1Ctx.Load(fileToUpload);
 
                     folder.EnsureFolder(TargetCopyFolderName);
 
+                    // Prereq for CopyFile_EmptyFolderBetweenSiteCollections_Test
+                    folder.EnsureFolder(EmptyFolderName);
+
+                    // Prereq for CopyFile_FolderWithSkipSourceFolderNameBetweenSiteCollections_Test
+                    var sourceFolder = folder.EnsureFolder(SourceFolderName);
+                    fileToUpload = sourceFolder.Files.Add(fci);
+                    site1Ctx.Load(fileToUpload);
+
+
+                    // Prereq for CopyFile_FolderWithFoldersAndEmptyFolderBetweenSiteCollections_Test
+                    var folderHirachyFolder0 = folder.EnsureFolder(SourceFolderWithFolders);
+                    var folderHirachyFolder1 = folderHirachyFolder0.EnsureFolder(TargetCopyFolderName);
+
+                    var folderHirachyFile0 = folderHirachyFolder0.Files.Add(fci);
+                    var folderHirachyFile1 = folderHirachyFolder1.Files.Add(fci);
+                    site1Ctx.Load(folderHirachyFile0);
+                    site1Ctx.Load(folderHirachyFile1);
+
+                    folderHirachyFolder1.EnsureFolder(EmptyFolderName);
+
+                    site1Ctx.ExecuteQueryRetry();
                 }
                 OfficeDevPnP.Core.Sites.SiteCollection.CreateAsync(ctx, new OfficeDevPnP.Core.Sites.CommunicationSiteCollectionCreationInformation()
                 {
@@ -144,7 +172,7 @@ namespace SharePointPnP.PowerShell.Tests
                 var results = scope.ExecuteCommand("Get-PnPFile",
                     new CommandParameter("Url", siteRelativeFileUrl),
                     new CommandParameter("AsFile"));
-                
+
                 Assert.IsFalse(results.Any());
             }
         }
@@ -158,7 +186,7 @@ namespace SharePointPnP.PowerShell.Tests
 
                 var results = scope.ExecuteCommand("Get-PnPFile",
                     new CommandParameter("Url", siteRelativeFileUrl));
-                
+
                 Assert.IsTrue(results.Any());
                 Assert.IsTrue(results[0].BaseObject.GetType() == typeof(Microsoft.SharePoint.Client.File));
             }
@@ -236,6 +264,33 @@ namespace SharePointPnP.PowerShell.Tests
         }
 
         [TestMethod]
+        public void CopyFile_WithHashtag_Test()
+        {
+            using (var scope = new PSTestScope(_site1Url, true))
+            {
+                var sourceUrl = $"{Site1RelativeFolderUrl}/{TargetFileNameWithHashtag}";
+                var destinationUrl = $"{Site1RelativeFolderUrl}/{TargetCopyFolderName}";
+                var destinationFileUrl = $"{destinationUrl}/{TargetFileNameWithHashtag}";
+
+                var results = scope.ExecuteCommand("Copy-PnPFile",
+                    new CommandParameter("SourceUrl", sourceUrl),
+                    new CommandParameter("TargetUrl", destinationUrl),
+                    new CommandParameter("Force"));
+
+                using (var ctx = TestCommon.CreateClientContext(_site1Url))
+                {
+                    File initialFile = ctx.Web.GetFileByServerRelativeUrl(destinationFileUrl);
+                    ctx.Load(initialFile);
+                    ctx.ExecuteQueryRetry();
+                    if (!initialFile.Exists)
+                    {
+                        Assert.Fail("Copied file cannot be found");
+                    }
+                }
+            }
+        }
+
+        [TestMethod]
         public void CopyFile_BetweenSiteCollections_Test()
         {
             using (var scope = new PSTestScope(_site1Url, true))
@@ -285,6 +340,135 @@ namespace SharePointPnP.PowerShell.Tests
                     {
                         Assert.Fail("Copied file cannot be found");
                     }
+                }
+            }
+        }
+
+        [TestMethod]
+        public void CopyFile_BetweenSiteCollectionsWithHashtag_Test()
+        {
+            using (var scope = new PSTestScope(_site1Url, true))
+            {
+                var sourceUrl = $"{Site1RelativeFolderUrl}/{TargetFileNameWithHashtag}";
+                var destinationFolderUrl = $"{Site2RelativeFolderUrl}";
+                string destinationFileUrl = $"{destinationFolderUrl}/{TargetFileNameWithHashtag}";
+
+                var results = scope.ExecuteCommand("Copy-PnPFile",
+                    new CommandParameter("SourceUrl", sourceUrl),
+                    new CommandParameter("TargetUrl", destinationFolderUrl),
+                    new CommandParameter("Force"));
+
+                using (var ctx = TestCommon.CreateClientContext(_site2Url))
+                {
+                    File initialFile = ctx.Web.GetFileByServerRelativeUrl(destinationFileUrl);
+                    ctx.Load(initialFile);
+                    ctx.ExecuteQuery();
+                    if (!initialFile.Exists)
+                    {
+                        Assert.Fail("Copied file cannot be found");
+                    }
+                }
+            }
+        }
+
+        [TestMethod]
+        public void CopyFile_EmptyFolder_Test()
+        {
+            using (var scope = new PSTestScope(_site1Url, true))
+            {
+                var sourceUrl = $"{Site1RelativeFolderUrl}/{EmptyFolderName}";
+                var destinationUrl = $"{Site1RelativeFolderUrl}/{TargetCopyFolderName}/{EmptyFolderName}";
+
+                var results = scope.ExecuteCommand("Copy-PnPFile",
+                    new CommandParameter("SourceUrl", sourceUrl),
+                    new CommandParameter("TargetUrl", destinationUrl),
+                    new CommandParameter("Force"));
+
+                using (var ctx = TestCommon.CreateClientContext(_site1Url))
+                {
+                    Folder initialFolder = ctx.Web.GetFolderByServerRelativePath(ResourcePath.FromDecodedUrl(destinationUrl));
+                    initialFolder.EnsureProperties(f => f.Name, f => f.Exists);
+                    ctx.Load(initialFolder);
+                    ctx.ExecuteQuery();
+                    if (!initialFolder.Exists)
+                    {
+                        Assert.Fail("Copied folder cannot be found");
+                    }
+                }
+            }
+        }
+
+        [TestMethod]
+        public void CopyFile_EmptyFolderBetweenSiteCollections_Test()
+        {
+            using (var scope = new PSTestScope(_site1Url, true))
+            {
+                var sourceUrl = $"{Site1RelativeFolderUrl}/{EmptyFolderName}";
+                var destinationUrl = $"{Site2RelativeFolderUrl}/{EmptyFolderName}";
+
+                var results = scope.ExecuteCommand("Copy-PnPFile",
+                    new CommandParameter("SourceUrl", sourceUrl),
+                    new CommandParameter("TargetUrl", destinationUrl),
+                    new CommandParameter("Force"));
+
+                using (var ctx = TestCommon.CreateClientContext(_site2Url))
+                {
+                    Folder initialFolder = ctx.Web.GetFolderByServerRelativePath(ResourcePath.FromDecodedUrl(destinationUrl));
+                    initialFolder.EnsureProperties(f => f.Name, f => f.Exists);
+                    ctx.Load(initialFolder);
+                    ctx.ExecuteQuery();
+                    if (!initialFolder.Exists)
+                    {
+                        Assert.Fail("Copied folder cannot be found");
+                    }
+                }
+            }
+        }
+
+        [TestMethod]
+        public void CopyFile_FolderWithSkipSourceFolderNameBetweenSiteCollections_Test()
+        {
+            using (var scope = new PSTestScope(_site1Url, true))
+            {
+                var sourceUrl = $"{Site1RelativeFolderUrl}/{SourceFolderName}";
+                var destinationUrl = $"{Site2RelativeFolderUrl}";
+
+                var results = scope.ExecuteCommand("Copy-PnPFile",
+                    new CommandParameter("SourceUrl", sourceUrl),
+                    new CommandParameter("TargetUrl", destinationUrl),
+                    new CommandParameter(name: "SkipSourceFolderName"),
+                    new CommandParameter("Force"));
+
+                using (var ctx = TestCommon.CreateClientContext(_site2Url))
+                {
+                    Folder initialFolder = ctx.Web.GetFolderByServerRelativePath(ResourcePath.FromDecodedUrl(destinationUrl));
+                    initialFolder.EnsureProperties(f => f.Name, f => f.Exists, f => f.Files);
+                    ctx.Load(initialFolder);
+                    ctx.ExecuteQuery();
+                    Assert.AreEqual(1, initialFolder.Files.Count);
+                }
+            }
+        }
+
+        [TestMethod]
+        public void CopyFile_FolderWithFoldersAndEmptyFolderBetweenSiteCollections_Test()
+        {
+            using (var scope = new PSTestScope(_site1Url, true))
+            {
+                var sourceUrl = $"{Site1RelativeFolderUrl}/{SourceFolderWithFolders}";
+                var destinationUrl = $"{Site2RelativeFolderUrl}";
+
+                var results = scope.ExecuteCommand("Copy-PnPFile",
+                    new CommandParameter("SourceUrl", sourceUrl),
+                    new CommandParameter("TargetUrl", destinationUrl),
+                    new CommandParameter("Force"));
+
+                using (var ctx = TestCommon.CreateClientContext(_site2Url))
+                {
+                    List list = ctx.Web.GetListUsingPath(ResourcePath.FromDecodedUrl(destinationUrl));
+                    ctx.Load(list);
+                    ctx.ExecuteQuery();
+                    Assert.AreEqual(5, list.ItemCount);
                 }
             }
         }