+++ CHANGELOG.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+
+### Added
+
+### Changed
+
+### Contributors
+
+## [3.14.1910.1]
+
+### Added
+
+- ConvertTo-PnPClientSidePage: Added support for logging to console via `-LogType Console`
+- Copy-PnPFile: Fixes (#2300)
+- ConvertTo-PnPClientSidePage: Added support for controlling the target page name is any cross site transformation (so wiki, web part, blog in addition the already existing option for publishing pages) via the `-TargetPageName` parameter
+
+### Changed
+
+### Contributors
+
 ## [3.14.1910.0]
 
 ### Added
@@ -18,6 +38,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 
+- Several documentation fixes
 - Add-PnPClientSideWebPart now also works for SP2019
 - Added -List parameter to Get-PnPFolder to retrieve all folders in a list
 - Added owner paramter to New-PnPSite when create Communications site
@@ -28,11 +49,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Contributors
 
+- Aleksandr SaPozhkov [shurick81]
 - Garry Trinder [garrytrinder]
 - Koen Zomers [KoenZomers]
 - Gautam Sheth [gautamdsheth]
 - Giacomo Pozzoni [jackpoz]
 - Paul Bullock [pkbullock]
+- Andres Mariano Gorzelany [get-itips]
 
 ## [3.13.1909.0]